feat:event-list-page-first-commit
Implemented a new component for the events list. It's likely that once the design is approved, this component will replace the event cards currently displayed on the homepage.

Integrated a package to detect the device type (mobile, tablet, or desktop) and adapt the experience accordingly.

Updated the main menu in both mobile and desktop versions to enhance navigation and user experience.

Modified the banner design for a more appealing and effective appearance.